# contracts/agent-internal-openapi.yaml
openapi: 3.1.0
info:
  title: Caijiatai Python Agent Internal API
  version: 0.5.0
  description: >-
    Internal-only API used by the Java procurement control plane.
    Production transport is Kafka RPC: commands arrive on topic caijiatai.commands as
    HMAC-SHA256 signed envelopes (MessageCodec.signEnvelope; signature over the
    canonical JSON excluding the signature field itself) and results return on
    caijiatai.results; both sides verify signature plus payload_sha256. The HTTP
    surface documented here is the legacy loopback mode: every /internal/v1/* path
    (and, in internal mode, every non-health path) is gated by the
    X-Agent-Internal-Token header with a constant-time compare; POST is only allowed
    for /internal/v1/commands and /internal/v1/config.
servers:
  - url: http://agent:8742
security:
  - InternalToken: []
paths:
...
  /api/runs/{runId}:
    get:
      operationId: getRun
      parameters: [{$ref: '#/components/parameters/RunId'}]
      responses:
        '200': {description: Run, content: {application/json: {schema: {type: object}}}}
  /api/runs/{runId}/report:
    get:
      operationId: getRunReport
      parameters: [{$ref: '#/components/parameters/RunId'}]
      responses:
        '200': {description: Runtime report, content: {application/json: {schema: {type: object}}}}
  /api/runs/{runId}/messages:
    get:
      operationId: getRunMessages
      parameters: [{$ref: '#/components/parameters/RunId'}]
      responses:
        '200': {description: Run messages, content: {application/json: {schema: {type: array, items: {type: object}}}}}
  /api/runs/{runId}/events:
    get:
      operationId: getRunEvents
      parameters: [{$ref: '#/components/parameters/RunId'}]
      responses:
        '200': {description: Run events, content: {application/json: {schema: {type: array, items: {type: object}}}}}
  /api/runs/{runId}/approvals:
    get:
      operationId: getRunApprovals
      parameters: [{$ref: '#/components/parameters/RunId'}]
      responses:
        '200': {description: Run approvals, content: {application/json: {schema: {type: array, items: {type: object}}}}}
  /api/runs/{runId}/tool-invocations:
    get:
      operationId: getRunToolInvocations
      parameters: [{$ref: '#/components/parameters/RunId'}]
      responses:
        '200': {description: Tool invocations, content: {application/json: {schema: {type: array, items: {type: object}}}}}
  /api/runs/{runId}/checkpoint:
    get:
      operationId: getRunCheckpoint
      parameters: [{$ref: '#/components/parameters/RunId'}]
      responses:
        '200': {description: Checkpoint, content: {application/json: {schema: {type: object}}}}
...
components:
  securitySchemes:
    InternalToken:
      type: apiKey
      in: header
      name: X-Agent-Internal-Token
      description: >-
        Legacy HTTP loopback mode only. In production Kafka RPC mode there is no
        HTTP hop between Java and Python; authenticity comes from the HMAC-SHA256
        envelope signature (shared internal HMAC key) plus payload_sha256 checks.
  parameters:
...
    RunId:
      name: runId
      in: path
      required: true
      schema: {type: string, pattern: '^[0-9a-f]{32}$'}
...